requirements.txt
openai>=1.0.0
pytesseract
pillow
PyMuPDF
pandas
fastapi
uvicorn
python-dotenv
